moved imgs to new folder; adjusted player creation message; got new pointer
