+++ src/Components/App/App.scss
@@ -6,4 +6,5 @@
 * {
   font-family: 'Indie Flower', cursive;
   font-weight: 600;
+  cursor: url("../../Components/Images/pointer.PNG"), auto;
 }
