+++ src/Components/GameBar/GameBar.jsx
@@ -1,19 +1,10 @@
 import React, {useEffect, useState} from 'react';
 import "./GameBar.scss";
-import logo from "../logo_concept.png";
+import logo from "../../Components/Images/logo_concept.png";
 import Player from "../Player/Player";
-import av from "../psy.png";
+import av from "../Images/psy.png";
 
-const GameBar = ({master, setMaster, party}) => {
-    const incrementMaster = () => {
-        const currentMasterValue = master;
-        setMaster(currentMasterValue + 1);
-    }
-
-    const decrementMaster = () => {
-        const currentMasterValue = master;
-        setMaster(currentMasterValue - 1);
-    }
+const GameBar = ({master, setMaster, party, incrementMaster, decrementMaster}) => {
 
     return (
         <>
