+++ src/Components/GameBar/GameBar.scss
@@ -26,7 +26,7 @@
     width: 90%;
     left: 50%;
     transform: translate(-50%, 0%);
-    cursor: pointer;
+    // cursor: pointer;
 }
 
 .partyContainer {
@@ -43,6 +43,3 @@
     width: 100%;
     height: 100%;
 }
-
-
-
