+++ src/Components/GameStart/GameStart.scss
@@ -19,7 +19,7 @@
 }
 
 .startButtonText {
-    cursor: pointer;
+    // cursor: pointer;
     padding: 5px;
     position: absolute;
     font-size: 1.5em;
