+++ src/Components/Inn/Inn.scss
@@ -1,5 +1,5 @@
 .hitbox {
-    cursor: pointer;
+    // cursor: pointer;
     // border: 1px solid;
     height: 100px;
     width: 100px;
