+++ src/Components/PlayerCreation/PlayerCreation.scss
@@ -81,7 +81,7 @@
     text-align: left;
     border-radius: 4px;
     margin: 10px 10px 20px 25px;
-    cursor: pointer;
+    // cursor: pointer;
     padding-bottom: 15px;
 }
 
@@ -137,7 +137,7 @@
     transform: translate(-50%, 0%);
     position: relative;
     text-align: center;
-    cursor: pointer;
+    // cursor: pointer;
     padding-top: 15px;
     // background-color: #d3edef;
     height: 12%;
